step3: move Input validator to view package
